requirements.txt
-i https://pypi.org/simple
asgiref==3.6.0; python_version >= '3.7'
boto3==1.26.143; python_version >= '3.7'
botocore==1.29.143; python_version >= '3.7'
django==4.1.5; python_version >= '3.8'
django-filter==22.1; python_version >= '3.7'
djangorestframework==3.14.0; python_version >= '3.6'
drf-extensions==0.7.1
jmespath==1.0.1; python_version >= '3.7'
pycodestyle==2.10.0; python_version >= '3.6'
pyjwt==2.6.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.7.1
s3transfer==0.6.1; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3; python_version >= '3.5'
unidecode==1.3.6; python_version >= '3.5'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
